implementing sendBufMsg and recvBufMsg.
