+++ Network/Socket/Posix/IOVec.hsc
@@ -3,7 +3,10 @@
 -- | Support module for the POSIX writev system call.
 module Network.Socket.Posix.IOVec
     ( IOVec(..)
+    , withIOVec
     ) where
+
+import Foreign.Marshal.Array (allocaArray)
 
 import Network.Socket.Imports
 
@@ -11,7 +14,7 @@ import Network.Socket.Imports
 #include <sys/uio.h>
 
 data IOVec = IOVec
-    { iovBase :: !(Ptr CChar)
+    { iovBase :: !(Ptr Word8)
     , iovLen  :: !CSize
     }
 
@@ -27,3 +30,17 @@ instance Storable IOVec where
   poke p iov = do
     (#poke struct iovec, iov_base) p (iovBase iov)
     (#poke struct iovec, iov_len)  p (iovLen  iov)
+
+-- | @withIOVec cs f@ executes the computation @f@, passing as argument a pair
+-- consisting of a pointer to a temporarily allocated array of pointers to
+-- IOVec made from @cs@ and the number of pointers (@length cs@).
+-- /Unix only/.
+withIOVec :: [(Ptr Word8, Int)] -> ((Ptr IOVec, Int) -> IO a) -> IO a
+withIOVec cs f =
+    allocaArray csLen $ \aPtr -> do
+        zipWithM_ pokeIov (ptrs aPtr) cs
+        f (aPtr, csLen)
+  where
+    csLen = length cs
+    ptrs = iterate (`plusPtr` sizeOf (IOVec nullPtr 0))
+    pokeIov ptr (sPtr, sLen) = poke ptr $ IOVec sPtr (fromIntegral sLen)
